requirements.txt
Pillow
torchvision
flask
lightning